requirements.txt
split-folders
opencv-python
numpy
tqdm
torch
torchvision
matplotlib
pandas
scikit-learn
Pillow